Move fixtures to conftest
We will need them in more modules of the test suite
